Refactored tests names
