+++ src/dcgSelenium/Tests/T_004_addBasicInfo.java
@@ -3,19 +3,23 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
+import dcgSelenium.Pages.PreviewPage;
 import dcgSelenium.Pages.SettingsPage;
 
-
-public class T_010 extends BaseTest {
+public class T_004_addBasicInfo extends BaseTest {
 
 	@Test
-	public void removeExperirence() {
+	public void addBasicInfo() {
+
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
+		PreviewPage previewPage = new PreviewPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
 		SettingsPage settingsPage = new SettingsPage(driver, wait);
-		settingsPage.removeExp();
-	}
+		settingsPage.fillBasicInfo();
+		homePage.goToPreview();
+		previewPage.checkBaiscInfo();
 
+	}
 }
