+++ src/dcgSelenium/Tests/T_005_addEducation.java
@@ -3,19 +3,21 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
+import dcgSelenium.Pages.PreviewPage;
 import dcgSelenium.Pages.SettingsPage;
 
-public class T_015 extends BaseTest {
+public class T_005_addEducation extends BaseTest {
 
 	@Test
-	public void newHire() {
-		
+	public void addEducation() {
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
+		PreviewPage previewPage = new PreviewPage(driver, wait);
+		SettingsPage settingsPage = new SettingsPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
-		SettingsPage settingsPage = new SettingsPage(driver, wait);
-		settingsPage.addNewHire();
+		settingsPage.fillEducation();
+		homePage.goToPreview();
+		previewPage.checkEducation();
 	}
-
 }
