+++ src/dcgSelenium/Tests/T_007_addSkillset.java
@@ -3,19 +3,23 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
+import dcgSelenium.Pages.PreviewPage;
 import dcgSelenium.Pages.SettingsPage;
 
-public class T_011 extends BaseTest {
+public class T_007_addSkillset extends BaseTest {
 
 	@Test
-	public void removeEducation() {
-		
+	public void addSkillset() {
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
+		PreviewPage previewPage = new PreviewPage(driver, wait);
+		SettingsPage settingsPage = new SettingsPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
-		SettingsPage settingsPage = new SettingsPage(driver, wait);
-		settingsPage.removeEdu();
+		settingsPage.addSkills();
+		homePage.goToPreview();
+		previewPage.checkSkills();
+		
 	}
 
 }
