+++ src/dcgSelenium/Tests/T_008_selectUser.java
@@ -3,18 +3,19 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
+import dcgSelenium.Pages.UsersPage;
 
-public class T_004 extends BaseTest {
+public class T_008_selectUser extends BaseTest {
 
 	@Test
-	public void logOut() {
-
+	public void selectUser() {
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("test@user.com", "asdasd1");
-		homePage.signOut();
-
+		UsersPage usersPage = new UsersPage(driver, wait);
+		homePage.goToUsers();
+		usersPage.selectUser();
 	}
 
 }
