+++ src/dcgSelenium/Tests/T_010_removeEducation.java
@@ -3,23 +3,19 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
-import dcgSelenium.Pages.PreviewPage;
 import dcgSelenium.Pages.SettingsPage;
 
-public class T_008 extends BaseTest {
+public class T_010_removeEducation extends BaseTest {
 
 	@Test
-	public void addSkillset() {
+	public void removeEducation() {
+		
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
-		PreviewPage previewPage = new PreviewPage(driver, wait);
-		SettingsPage settingsPage = new SettingsPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
-		settingsPage.addSkills();
-		homePage.goToPreview();
-		previewPage.checkSkills();
-		
+		SettingsPage settingsPage = new SettingsPage(driver, wait);
+		settingsPage.removeEdu();
 	}
 
 }
