+++ src/dcgSelenium/Tests/T_011_removeSkills.java
@@ -5,18 +5,17 @@ import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
 import dcgSelenium.Pages.SettingsPage;
 
-public class T_014 extends BaseTest {
+public class T_011_removeSkills extends BaseTest {
 
 	@Test
-	public void removeMainProject() {
+	public void removeSkills() {
 		
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
 		SettingsPage settingsPage = new SettingsPage(driver, wait);
-		settingsPage.removeMainProj();
-		homePage.goToPreview();
+		settingsPage.removeSkll();
 	}
 
 }
