+++ src/dcgSelenium/Tests/T_012_removeProject.java
@@ -5,17 +5,18 @@ import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
 import dcgSelenium.Pages.SettingsPage;
 
-public class T_012 extends BaseTest {
+public class T_012_removeProject extends BaseTest {
 
 	@Test
-	public void removeSkills() {
+	public void removeProject() {
 		
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
 		homePage.logInUrl();
 		loginPage.LogIn("newuser2@ds.com", "pass123");
 		SettingsPage settingsPage = new SettingsPage(driver, wait);
-		settingsPage.removeSkll();
+		settingsPage.removeProj();
+		homePage.goToPreview();
 	}
 
 }
