+++ src/dcgSelenium/Tests/T_014_newHire.java
@@ -3,19 +3,19 @@ package dcgSelenium.Tests;
 import org.testng.annotations.Test;
 import dcgSelenium.Pages.HomePage;
 import dcgSelenium.Pages.LoginPage;
-import dcgSelenium.Pages.UsersPage;
+import dcgSelenium.Pages.SettingsPage;
 
-public class T_009 extends BaseTest {
+public class T_014_newHire extends BaseTest {
 
 	@Test
-	public void selectUser() {
+	public void newHire() {
+		
 		HomePage homePage = new HomePage(driver, wait);
 		LoginPage loginPage = new LoginPage(driver, wait);
 		homePage.logInUrl();
-		loginPage.LogIn("test@user.com", "asdasd1");
-		UsersPage usersPage = new UsersPage(driver, wait);
-		homePage.goToUsers();
-		usersPage.selectUser();
+		loginPage.LogIn("newuser2@ds.com", "pass123");
+		SettingsPage settingsPage = new SettingsPage(driver, wait);
+		settingsPage.addNewHire();
 	}
 
 }
